+++ tests/ttnn/unit_tests/gtests/sources.cmake
@@ -17,7 +17,6 @@ set(UNIT_TESTS_TTNN_BASIC_SOURCES
     test_broadcast_to.cpp
     test_convert_to_hwc_gather.cpp
     test_gelu_bw_ulp.cpp
-    test_gelu_bw_main_ulp.cpp
     test_generic_op.cpp
     test_generic_op_hashing.cpp
     test_graph_add.cpp
