--- tests/ttnn/unit_tests/operations/eltwise/test_gelu_bw_main_ulp.py
@@ -1,469 +1,0 @@
-# SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
-#
-# SPDX-License-Identifier: Apache-2.0
-
-"""
-GELU Backward ULP Precision Tests
-
-This test validates the accuracy of ttnn.gelu_bw (GELU derivative) across
-the BFloat16 range using the same methodology as test_gelu_floor_value_bug.py.
-
-MATHEMATICAL FORMULA:
-GELU'(x) = grad * (cdf + x * pdf)
-where:
-  cdf = 0.5 * (1 + erf(x / sqrt(2)))  -- CDF of standard normal distribution
-  pdf = exp(-x^2 / 2) / sqrt(2*pi)    -- PDF of standard normal distribution
-
-Hardware Model: Tenstorrent SFPU uses DAZ+FTZ (Denormals-Are-Zero + Flush-To-Zero)
-Per tech_reports/Handling_Special_Value/special_values.md: "denormals | all | 0x0"
-
-Run: pytest tests/ttnn/unit_tests/operations/eltwise/test_gelu_bw_main_ulp.py -v -s
-"""
-
-import pytest
-import torch
-import ttnn
-from loguru import logger
-from mpmath import mp, erf as mp_erf, erfc as mp_erfc, exp as mp_exp, sqrt as mp_sqrt
-from tests.ttnn.unit_tests.operations.eltwise.eltwise_test_utils import (
-    float_to_bf16_bits,
-    bf16_bits_to_float,
-    bf16_daz_normalize,
-    ulp_distance_bf16_daz,
-)
-
-
-def gelu_derivative_exact(x: float) -> float:
-    """
-    Exact GELU derivative using mpmath 256-bit precision.
-
-    GELU'(x) = cdf + x * pdf
-    where:
-      cdf = 0.5 * (1 + erf(x / sqrt(2)))
-      pdf = exp(-x^2 / 2) / sqrt(2*pi)
-
-    Uses erfc() for numerical stability with large negative x.
-    """
-    mp.prec = 256
-    x_mp = mp.mpf(x)
-    sqrt2 = mp_sqrt(2)
-    sqrt_2pi = mp_sqrt(2 * mp.pi)
-
-    # CDF: 0.5 * (1 + erf(x / sqrt(2)))
-    # For numerical stability with negative x, use erfc
-    if x < 0:
-        cdf = mp.mpf("0.5") * mp_erfc(-x_mp / sqrt2)
-    else:
-        cdf = mp.mpf("0.5") * (1 + mp_erf(x_mp / sqrt2))
-
-    # PDF: exp(-x^2 / 2) / sqrt(2*pi)
-    pdf = mp_exp(-x_mp * x_mp / 2) / sqrt_2pi
-
-    # GELU'(x) = cdf + x * pdf
-    result = cdf + x_mp * pdf
-    return float(result)
-
-
-def gelu_derivative_expected_bf16_daz(x: float) -> float:
-    """Compute expected BF16 GELU derivative with DAZ+FTZ applied."""
-    x_bits = bf16_daz_normalize(float_to_bf16_bits(x))
-    x_daz = bf16_bits_to_float(x_bits)
-    result = gelu_derivative_exact(x_daz)
-    result_bits = bf16_daz_normalize(float_to_bf16_bits(result))
-    return bf16_bits_to_float(result_bits)
-
-
-def gelu_bw_expected_bf16_daz(grad: float, x: float) -> float:
-    """Compute expected BF16 GELU backward with DAZ+FTZ applied."""
-    grad_bits = bf16_daz_normalize(float_to_bf16_bits(grad))
-    grad_daz = bf16_bits_to_float(grad_bits)
-    x_bits = bf16_daz_normalize(float_to_bf16_bits(x))
-    x_daz = bf16_bits_to_float(x_bits)
-    result = grad_daz * gelu_derivative_exact(x_daz)
-    result_bits = bf16_daz_normalize(float_to_bf16_bits(result))
-    return bf16_bits_to_float(result_bits)
-
-
-# =============================================================================
-# Test Classes
-# =============================================================================
-
-
-class TestGeluBwDerivativeAtZero:
-    """Correctness guard: verifies GELU'(0) = 0.5 via full Python → TTNN → device round-trip.
-    Catches kernel crash, wrong formula sign, or broken CDF constant."""
-
-    def test_derivative_at_zero(self, device):
-        """GELU'(0) = 0.5"""
-        input_val = 0.0
-        grad_val = 1.0
-
-        torch_input = torch.tensor([[input_val]], dtype=torch.bfloat16)
-        torch_grad = torch.tensor([[grad_val]], dtype=torch.bfloat16)
-
-        tt_input = ttnn.from_torch(torch_input, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-        tt_grad = ttnn.from_torch(torch_grad, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-
-        results = ttnn.gelu_bw(tt_grad, tt_input, approximate="none")
-        actual = ttnn.to_torch(results[0]).item()
-
-        expected = gelu_derivative_expected_bf16_daz(input_val)
-        ulp_error = ulp_distance_bf16_daz(actual, expected)
-
-        logger.debug(f"x=0: expected={expected:.4f}, actual={actual:.4f}, ULP={ulp_error}")
-
-        assert ulp_error <= 2, f"Expected ULP <= 2, got {ulp_error}"
-
-
-class TestGeluBwPositiveValues:
-    """Correctness guard: 6 positive-side points where GELU'(x) approaches 1.
-    Per-point ULP threshold (2). Catches broken positive saturation path."""
-
-    @pytest.mark.parametrize(
-        "input_value,max_expected_ulp",
-        [
-            (0.5, 2),
-            (1.0, 2),
-            (2.0, 2),
-            (3.0, 2),
-            (5.0, 2),
-            (10.0, 2),
-        ],
-    )
-    def test_positive_values(self, device, input_value, max_expected_ulp):
-        """For large positive x, GELU'(x) approaches 1."""
-        torch_input = torch.tensor([[input_value]], dtype=torch.bfloat16)
-        torch_grad = torch.tensor([[1.0]], dtype=torch.bfloat16)
-
-        tt_input = ttnn.from_torch(torch_input, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-        tt_grad = ttnn.from_torch(torch_grad, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-
-        results = ttnn.gelu_bw(tt_grad, tt_input, approximate="none")
-        actual = ttnn.to_torch(results[0]).item()
-
-        expected = gelu_derivative_expected_bf16_daz(input_value)
-        ulp_error = ulp_distance_bf16_daz(actual, expected)
-
-        logger.debug(f"x={input_value}: expected={expected:.4f}, actual={actual:.4f}, ULP={ulp_error}")
-
-        assert ulp_error <= max_expected_ulp, f"Expected ULP <= {max_expected_ulp}, got {ulp_error}"
-
-
-class TestGeluBwNegativeValues:
-    """Correctness guard: 8 negative-side points with tight per-point thresholds (2 ULP).
-    Catches precision regression at any negative sample point."""
-
-    @pytest.mark.parametrize(
-        "input_value,max_expected_ulp",
-        [
-            (-0.5, 2),
-            (-1.0, 2),
-            (-2.0, 2),
-            (-3.0, 2),
-            (-4.0, 2),
-            (-5.0, 2),
-            (-6.0, 2),
-            (-8.0, 2),
-        ],
-    )
-    def test_negative_values(self, device, input_value, max_expected_ulp):
-        """For large negative x, GELU'(x) approaches 0."""
-        torch_input = torch.tensor([[input_value]], dtype=torch.bfloat16)
-        torch_grad = torch.tensor([[1.0]], dtype=torch.bfloat16)
-
-        tt_input = ttnn.from_torch(torch_input, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-        tt_grad = ttnn.from_torch(torch_grad, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-
-        results = ttnn.gelu_bw(tt_grad, tt_input, approximate="none")
-        actual = ttnn.to_torch(results[0]).item()
-
-        expected = gelu_derivative_expected_bf16_daz(input_value)
-        ulp_error = ulp_distance_bf16_daz(actual, expected)
-
-        logger.debug(f"x={input_value}: expected={expected:.6e}, actual={actual:.6e}, ULP={ulp_error}")
-
-        assert ulp_error <= max_expected_ulp, f"Expected ULP <= {max_expected_ulp}, got {ulp_error}"
-
-
-class TestGeluBwNearZero:
-    """Correctness guard: 7 near-zero points where GELU'(x) ≈ 0.5 + x/sqrt(2π).
-    Tight threshold (2 ULP). Catches DAZ flush bugs or broken small-value handling."""
-
-    @pytest.mark.parametrize(
-        "input_value",
-        [1e-6, 1e-4, 0.01, 0.1, -0.1, -0.01, -1e-4],
-    )
-    def test_near_zero(self, device, input_value):
-        """Near zero, GELU'(x) ≈ 0.5 + x/sqrt(2*pi)."""
-        torch_input = torch.tensor([[input_value]], dtype=torch.bfloat16)
-        torch_grad = torch.tensor([[1.0]], dtype=torch.bfloat16)
-
-        tt_input = ttnn.from_torch(torch_input, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-        tt_grad = ttnn.from_torch(torch_grad, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-
-        results = ttnn.gelu_bw(tt_grad, tt_input, approximate="none")
-        actual = ttnn.to_torch(results[0]).item()
-
-        expected = gelu_derivative_expected_bf16_daz(input_value)
-        ulp_error = ulp_distance_bf16_daz(actual, expected)
-
-        logger.debug(f"x={input_value:.2e}: expected={expected:.4f}, actual={actual:.4f}, ULP={ulp_error}")
-
-        assert ulp_error <= 2, f"Expected ULP <= 2, got {ulp_error}"
-
-
-class TestGeluBwLocalMinimum:
-    """Correctness guard: 5 points near the GELU derivative's zero-crossing at
-    x ≈ -0.751. Uses absolute error threshold (0.01) because near the zero-crossing,
-    the expected value is near 0 and even tiny absolute errors produce large ULPs."""
-
-    @pytest.mark.parametrize(
-        "input_value",
-        [-0.7, -0.75, -0.751, -0.76, -0.8],
-    )
-    def test_local_minimum_region(self, device, input_value):
-        """GELU has a local minimum around x ≈ -0.751 where GELU'(x) ≈ 0."""
-        torch_input = torch.tensor([[input_value]], dtype=torch.bfloat16)
-        torch_grad = torch.tensor([[1.0]], dtype=torch.bfloat16)
-
-        tt_input = ttnn.from_torch(torch_input, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-        tt_grad = ttnn.from_torch(torch_grad, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-
-        results = ttnn.gelu_bw(tt_grad, tt_input, approximate="none")
-        actual = ttnn.to_torch(results[0]).item()
-
-        expected = gelu_derivative_expected_bf16_daz(input_value)
-        ulp_error = ulp_distance_bf16_daz(actual, expected)
-
-        logger.debug(f"x={input_value}: expected={expected:.6f}, actual={actual:.6f}, ULP={ulp_error}")
-
-        # Near the zero-crossing, expected ≈ 0 so ULP can be large even for tiny absolute errors
-        assert (
-            ulp_error <= 2 or abs(actual - expected) < 0.01
-        ), f"ULP {ulp_error} and abs error {abs(actual - expected):.6f} both exceed thresholds"
-
-
-class TestGeluBwWithGradientScaling:
-    """Correctness guard (unique): only tests using grad != 1.0 (grad=0.5, 2.0, -1.0, 0.1, 10.0).
-    Catches swapped grad/input tensors or missing gradient multiplication in backward pass."""
-
-    @pytest.mark.parametrize(
-        "input_value,grad_value,max_expected_ulp",
-        [
-            (1.0, 2.0, 2),
-            (-1.0, 0.5, 2),
-            (0.0, 1.0, 2),
-            (2.0, -1.0, 2),
-            (0.5, 3.0, 2),
-        ],
-    )
-    def test_with_gradient(self, device, input_value, grad_value, max_expected_ulp):
-        """Test GELU backward with different gradient values."""
-        torch_input = torch.tensor([[input_value]], dtype=torch.bfloat16)
-        torch_grad = torch.tensor([[grad_value]], dtype=torch.bfloat16)
-
-        tt_input = ttnn.from_torch(torch_input, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-        tt_grad = ttnn.from_torch(torch_grad, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-
-        results = ttnn.gelu_bw(tt_grad, tt_input, approximate="none")
-        actual = ttnn.to_torch(results[0]).item()
-
-        expected = gelu_bw_expected_bf16_daz(grad_value, input_value)
-        ulp_error = ulp_distance_bf16_daz(actual, expected)
-
-        logger.debug(
-            f"x={input_value}, grad={grad_value}: expected={expected:.4f}, actual={actual:.4f}, ULP={ulp_error}"
-        )
-
-        assert ulp_error <= max_expected_ulp, f"Expected ULP <= {max_expected_ulp}, got {ulp_error}"
-
-
-def test_gelu_bw_ulp_summary(device):
-    """Correctness guard: comprehensive summary of GELU backward ULP across all regions.
-    Tests 6 key points (zero, ±1, local min, ±5) and asserts max ULP <= 2.
-    Catches any single broken code path quickly via representative sampling."""
-    logger.debug("")
-    logger.debug("=" * 100)
-    logger.debug("GELU BACKWARD ULP SUMMARY (DAZ+FTZ MODEL)")
-    logger.debug("=" * 100)
-
-    # Key test points
-    test_points = [
-        ("Zero", 0.0),
-        ("Small positive", 0.1),
-        ("Unity", 1.0),
-        ("Moderate positive", 2.0),
-        ("Large positive", 5.0),
-        ("Small negative", -0.1),
-        ("Negative unity", -1.0),
-        ("Local minimum", -0.751),
-        ("Moderate negative", -2.0),
-        ("Large negative", -5.0),
-        ("Deep negative", -8.0),
-    ]
-
-    logger.debug("")
-    logger.debug(f"{'Description':>20} | {'x':>10} | {'Expected':>12} | {'Actual':>12} | {'ULP':>8}")
-    logger.debug("-" * 70)
-
-    max_ulp = 0
-    worst_x = 0
-
-    for desc, x in test_points:
-        torch_input = torch.tensor([[x]], dtype=torch.bfloat16)
-        torch_grad = torch.tensor([[1.0]], dtype=torch.bfloat16)
-
-        tt_input = ttnn.from_torch(torch_input, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-        tt_grad = ttnn.from_torch(torch_grad, dtype=ttnn.bfloat16, device=device, layout=ttnn.TILE_LAYOUT)
-
-        results = ttnn.gelu_bw(tt_grad, tt_input, approximate="none")
-        actual = ttnn.to_torch(results[0]).item()
-
-        expected = gelu_derivative_expected_bf16_daz(x)
-        ulp = ulp_distance_bf16_daz(actual, expected)
-
-        if ulp > max_ulp:
-            max_ulp = ulp
-            worst_x = x
-
-        logger.debug(f"{desc:>20} | {x:>10.4f} | {expected:>12.6f} | {actual:>12.6f} | {ulp:>8}")
-
-    logger.debug("-" * 70)
-    logger.debug(f"Max ULP: {max_ulp} at x = {worst_x}")
-    logger.debug("")
-    logger.debug("Expected behavior:")
-    logger.debug("- GELU'(0) ≈ 0.5")
-    logger.debug("- GELU'(x) → 1 as x → +∞")
-    logger.debug("- GELU'(x) → 0 as x → -∞")
-    logger.debug("- Local minimum near x ≈ -0.751 where GELU'(x) ≈ 0")
-    logger.debug("=" * 100)
-
-    # Regression guard: all key points must be within 5 ULP
-    assert max_ulp <= 2, (
-        f"Max ULP {max_ulp} at x={worst_x} exceeds threshold 2. " f"See table above for per-point details."
-    )
-
-
-# =============================================================================
-# Gelu_bw Tanh approximation Exhaustive BF16 ULP Distribution Test
-# =============================================================================
-
-
-def test_gelu_bw_bf16_exhaustive(device):
-    """Exhaustive ULP distribution test for gelu_bw (approximate='tanh') across all BF16 values.
-
-    Generates all valid BF16 bit patterns as input, uses grad=1.0, and measures
-    ULP distance between device output and PyTorch float32 reference (tanh-approximate GELU derivative).
-    """
-    # Generate all bf16 bit patterns
-    all_bitpatterns = torch.arange(0, 2**16, dtype=torch.int32).to(torch.uint16)
-    vals = all_bitpatterns.view(torch.bfloat16)
-
-    # Filter to finite, non-zero, non-subnormal values
-    min_abs = torch.finfo(torch.bfloat16).tiny
-    max_abs = torch.finfo(torch.bfloat16).max
-    vals_f32_abs = vals.to(torch.float32).abs()
-    mask = torch.isfinite(vals) & (vals_f32_abs >= min_abs) & (vals_f32_abs <= max_abs) & (vals_f32_abs != 0)
-
-    value_set = vals[mask]
-    N = value_set.numel()
-    logger.debug(
-        f"Testing gelu_bw (approximate=tanh) with {N} BF16 values in [{value_set.min().item():.2e}, {value_set.max().item():.2e}]"
-    )
-
-    # Pad to multiple of 32 for tile layout
-    pad_size = (32 - (N % 32)) % 32
-    if pad_size > 0:
-        value_set_padded = torch.cat([value_set, torch.zeros(pad_size, dtype=torch.bfloat16)])
-    else:
-        value_set_padded = value_set
-
-    total_padded = value_set_padded.numel()
-    value_set_2d = value_set_padded.reshape(1, total_padded)
-
-    # Compute reference: GELU derivative via PyTorch autograd in float32
-    x_f32 = value_set_2d.to(torch.float32).requires_grad_(True)
-    y = torch.nn.functional.gelu(x_f32, approximate="tanh")
-    y.backward(torch.ones_like(y))
-    z_torch = x_f32.grad.detach()
-
-    # Run on device: gelu_bw with grad=1.0
-    grad_2d = torch.ones_like(value_set_2d)
-    tt_input = ttnn.from_torch(value_set_2d, dtype=ttnn.bfloat16, layout=ttnn.TILE_LAYOUT, device=device)
-    tt_grad = ttnn.from_torch(grad_2d, dtype=ttnn.bfloat16, layout=ttnn.TILE_LAYOUT, device=device)
-
-    results = ttnn.gelu_bw(tt_grad, tt_input, approximate="tanh")
-    tt_out = ttnn.to_torch(results[0])
-
-    # Trim padding
-    z_torch = z_torch[:, :N]
-    tt_out = tt_out[:, :N]
-
-    # Filter out inf/nan results
-    valid_mask = torch.isfinite(z_torch)
-    assert torch.isfinite(
-        tt_out[valid_mask]
-    ).all(), "ttnn.gelu_bw(approximate='tanh') output is non-finite where the reference is finite"
-    z_torch_valid = z_torch[valid_mask]
-    tt_out_valid = tt_out[valid_mask]
-    N_valid = z_torch_valid.numel()
-
-    logger.debug(f"Valid (finite) outputs: {N_valid}/{N} ({N_valid/N*100:.2f}%)")
-
-    # ULP computation in BF16 space
-    z_bf16 = z_torch_valid.to(torch.bfloat16).contiguous()
-    tt_bf16 = tt_out_valid.to(torch.bfloat16).contiguous()
-
-    # Flush subnormals and inf/max-normal to zero (DAZ+FTZ model)
-    z_bits_u16 = z_bf16.view(torch.uint16).to(torch.int32)
-    tt_bits_u16 = tt_bf16.view(torch.uint16).to(torch.int32)
-    subnormal_z = ((z_bits_u16 & 0x7F80) == 0) & ((z_bits_u16 & 0x7F) != 0)
-    subnormal_tt = ((tt_bits_u16 & 0x7F80) == 0) & ((tt_bits_u16 & 0x7F) != 0)
-    max_or_inf_z = (z_bits_u16 & 0x7F80) == 0x7F80
-    max_or_inf_tt = (tt_bits_u16 & 0x7F80) == 0x7F80
-    flush_mask = subnormal_z | subnormal_tt | max_or_inf_z | max_or_inf_tt
-    z_bf16 = torch.where(flush_mask, torch.zeros_like(z_bf16), z_bf16)
-    tt_bf16 = torch.where(flush_mask, torch.zeros_like(tt_bf16), tt_bf16)
-
-    z_bits = z_bf16.view(torch.uint16).to(torch.int32)
-    tt_bits = tt_bf16.view(torch.uint16).to(torch.int32)
-
-    sign_z = (z_bits >> 15) & 1
-    sign_tt = (tt_bits >> 15) & 1
-    z_ord = torch.clamp((z_bits & 0x7FFF) - 0x7F, min=0) * torch.where(sign_z == 0, 1, -1)
-    tt_ord = torch.clamp((tt_bits & 0x7FFF) - 0x7F, min=0) * torch.where(sign_tt == 0, 1, -1)
-    ulp_dist = (z_ord - tt_ord).abs()
-
-    max_ulp = ulp_dist.max().item()
-
-    # ULP distribution
-    ulp_0_count = (ulp_dist == 0).sum().item()
-    ulp_1_count = (ulp_dist == 1).sum().item()
-    ulp_2_count = (ulp_dist == 2).sum().item()
-    ulp_3_to_10_count = ((ulp_dist >= 3) & (ulp_dist <= 10)).sum().item()
-    ulp_11_to_100_count = ((ulp_dist >= 11) & (ulp_dist <= 100)).sum().item()
-    ulp_above_100_count = (ulp_dist > 100).sum().item()
-
-    mismatch_threshold = 2
-    mismatch_mask = ulp_dist > mismatch_threshold
-    total_mismatches = mismatch_mask.sum().item()
-    mismatch_pct = (total_mismatches / N_valid) * 100 if N_valid > 0 else 0.0
-
-    logger.debug(
-        f"Max ULP: {max_ulp}, mismatches (ULP > {mismatch_threshold}): {total_mismatches}/{N_valid} ({mismatch_pct:.4f}%)"
-    )
-    logger.debug(f"\nULP Distribution:")
-    logger.debug(f"  ULP = 0: {ulp_0_count:,} ({ulp_0_count/N_valid*100:.4f}%)")
-    logger.debug(f"  ULP = 1: {ulp_1_count:,} ({ulp_1_count/N_valid*100:.4f}%)")
-    logger.debug(f"  ULP = 2: {ulp_2_count:,} ({ulp_2_count/N_valid*100:.4f}%)")
-    logger.debug(f"  ULP 3-10: {ulp_3_to_10_count:,} ({ulp_3_to_10_count/N_valid*100:.4f}%)")
-    logger.debug(f"  ULP 11-100: {ulp_11_to_100_count:,} ({ulp_11_to_100_count/N_valid*100:.4f}%)")
-    logger.debug(f"  ULP > 100: {ulp_above_100_count:,} ({ulp_above_100_count/N_valid*100:.4f}%)")
-
-    # Verify counts sum correctly
-    ulp_sum = ulp_0_count + ulp_1_count + ulp_2_count + ulp_3_to_10_count + ulp_11_to_100_count + ulp_above_100_count
-    assert ulp_sum == N_valid, f"ULP counts don't sum to total valid: {ulp_sum} != {N_valid}"
-    # Compare device output against the float32 reference. Upcast the bf16 device output to float32 (lossless)
-    assert torch.allclose(
-        z_torch_valid, tt_out_valid.to(torch.float32), rtol=2e-2, atol=2e-2
-    ), "gelu_bw(approximate='tanh') output does not match float32 reference within tolerance"
