--- ttnn/cpp/ttnn/operations/experimental/unary_backward/gelu_backward/device/gelu_backward_device_operation.hpp
@@ -1,43 +1,0 @@
-// SPDX-FileCopyrightText: © 2025 Tenstorrent USA, Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-#include <optional>
-#include <string>
-#include <variant>
-
-#include "ttnn/device_operation.hpp"
-#include "ttnn/tensor/tensor.hpp"
-#include "gelu_backward_program_factory.hpp"
-#include "gelu_backward_device_operation_types.hpp"
-
-namespace ttnn::experimental::prim {
-
-struct GeluBackwardDeviceOperation {
-    using operation_attributes_t = GeluBackwardParams;
-    using tensor_args_t = GeluBackwardInputs;
-    using spec_return_value_t = tt::tt_metal::TensorSpec;
-    using tensor_return_value_t = Tensor;
-    using program_factory_t = std::variant<GeluBackwardProgramFactory>;
-
-    static void validate_on_program_cache_miss(const operation_attributes_t&, const tensor_args_t&);
-
-    static spec_return_value_t compute_output_specs(const operation_attributes_t&, const tensor_args_t&);
-    static tensor_return_value_t create_output_tensors(const operation_attributes_t& args, const tensor_args_t&);
-};
-
-}  // namespace ttnn::experimental::prim
-
-namespace ttnn::prim {
-
-Tensor gelu_bw(
-    const Tensor& grad_output,
-    const Tensor& input,
-    const std::string& approximate,
-    DataType output_dtype,
-    const MemoryConfig& output_memory_config,
-    const std::optional<Tensor>& preallocated_output);
-
-}  // namespace ttnn::prim
