--- ttnn/cpp/ttnn/operations/experimental/unary_backward/gelu_backward/device/gelu_backward_device_operation_types.hpp
@@ -1,24 +1,0 @@
-// SPDX-FileCopyrightText: © 2025 Tenstorrent USA, Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-#include <optional>
-#include "ttnn/tensor/tensor.hpp"
-
-namespace ttnn::experimental::prim {
-
-struct GeluBackwardParams {
-    const tt::tt_metal::DataType output_dtype = tt::tt_metal::DataType::INVALID;
-    const tt::tt_metal::MemoryConfig output_memory_config;
-    const std::string approximate = "none";
-};
-
-struct GeluBackwardInputs {
-    const Tensor& grad_output;
-    const Tensor& input;
-    std::optional<Tensor> preallocated_input_grad;
-};
-
-}  // namespace ttnn::experimental::prim
