--- ttnn/cpp/ttnn/operations/experimental/unary_backward/gelu_backward/gelu_backward.hpp
@@ -1,19 +1,0 @@
-// SPDX-FileCopyrightText: © 2025 Tenstorrent USA, Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-#include "ttnn/core.hpp"
-#include "ttnn/types.hpp"
-
-namespace ttnn::experimental {
-
-Tensor gelu_bw(
-    const Tensor& grad_output_tensor,
-    const Tensor& input_tensor,
-    const std::string& approximate = "none",
-    const std::optional<MemoryConfig>& memory_config = std::nullopt,
-    std::optional<Tensor> input_grad_tensor = std::nullopt);
-
-}  // namespace ttnn::experimental
